#493 added all unit tests for thermal submodels
