+++ pybamm/models/submodels/thermal/xyz_lumped/xyz_lumped_1D_current_collector.py
