+++ pybamm/models/submodels/thermal/xyz_lumped/xyz_lumped_2D_current_collector.py
