--- tests/unit/test_models/test_submodels/test_thermal/test_isothermal.py
@@ -1,25 +1,0 @@
-#
-# Test isothermal submodel
-#
-
-import pybamm
-import tests
-import unittest
-
-
-class TestIsothermal(unittest.TestCase):
-    def test_public_functions(self):
-        param = pybamm.standard_parameters_lithium_ion
-        submodel = pybamm.thermal.Isothermal(param)
-        std_tests = tests.StandardSubModelTests(submodel)
-        std_tests.test_all()
-
-
-if __name__ == "__main__":
-    print("Add -v for more debug output")
-    import sys
-
-    if "-v" in sys.argv:
-        debug = True
-    pybamm.settings.debug_mode = True
-    unittest.main()
