+++ tests/unit/test_models/test_submodels/test_thermal/test_xyz_lumped/test_base_xyz_lumped.py
@@ -1,5 +1,5 @@
 #
-# Test base thermal submodel
+# Test base xyz-lumped thermal submodel
 #
 
 import pybamm
@@ -9,7 +9,7 @@ import unittest
 
 class TestBaseModel(unittest.TestCase):
     def test_public_functions(self):
-        submodel = pybamm.thermal.current_collector.BaseModel(None)
+        submodel = pybamm.thermal.xyz_lumped.BaseModel(None)
         std_tests = tests.StandardSubModelTests(submodel)
         with self.assertRaises(NotImplementedError):
             std_tests.test_all()
@@ -23,3 +23,4 @@ if __name__ == "__main__":
         debug = True
     pybamm.settings.debug_mode = True
     unittest.main()
+
