+++ tests/unit/test_models/test_submodels/test_thermal/test_xyz_lumped/test_xyz_lumped_2D_current_collector.py
@@ -1,24 +1,25 @@
 #
-# Test isothermal submodel
+# Test xyz-lumped submodel with 2D current collectors
 #
 
 import pybamm
 import tests
 import unittest
 
+from tests.unit.test_models.test_submodels.test_thermal.coupled_variables import (
+    coupled_variables,
+)
 
-class TestIsothermal(unittest.TestCase):
+
+class TestCurrentCollector2D(unittest.TestCase):
     def test_public_functions(self):
         param = pybamm.standard_parameters_lithium_ion
         i_boundary_cc = pybamm.PrimaryBroadcast(pybamm.Scalar(1), ["current collector"])
-        variables = {"Current collector current density": i_boundary_cc}
 
-        submodel = pybamm.thermal.current_collector.Isothermal2D(param)
-        std_tests = tests.StandardSubModelTests(submodel, variables)
-        std_tests.test_all()
+        coupled_variables.update({"Current collector current density": i_boundary_cc})
 
-        submodel = pybamm.thermal.current_collector.Isothermal3D(param)
-        std_tests = tests.StandardSubModelTests(submodel, variables)
+        submodel = pybamm.thermal.xyz_lumped.CurrentCollector2D(param)
+        std_tests = tests.StandardSubModelTests(submodel, coupled_variables)
         std_tests.test_all()
